+++ Peng/RendererImplementation.h
@@ -6,7 +6,7 @@
 namespace Peng {
 	class RendererImplementation {
 	public:
-		virtual void Draw(Peng::Sprite& picture, int xPos, int yPos, int zPos, Peng::Shader shader) = 0;
+		virtual void Draw(Peng::Sprite& picture, float xPos, float yPos, float zPos, Peng::Shader shader) = 0;
 		virtual void ClearScreen() = 0;
 	};
 }
