+++ Peng/specificOpenGL/OpenGLRenderer.h
@@ -6,7 +6,7 @@
 
 namespace Peng {
 	class OpenGLRenderer : public RendererImplementation {
-		virtual void Draw(Peng::Sprite& picture, int xPos, int yPos, int zPos, Peng::Shader shader) override;
+		virtual void Draw(Peng::Sprite& picture, float xPos, float yPos, float zPos, Peng::Shader shader) override;
 		virtual void ClearScreen() override;
 
 	};
